use int8 for target

+++ e2e_recsys/data_generation/file_converter.py
@@ -1,5 +1,6 @@
 import os
 from typing import List, Optional
+import numpy as np
 import pandas as pd
 import torch
 
@@ -48,7 +49,11 @@ class FileConverter:
         row_data = next(self.data_loader)
         for col in self.columns:
             row_dict[col] = row_data[col].to_numpy()[0]
-        self.row = (row_dict, row_data[self.target_col].to_numpy()[0])
+        # For binary classification, set target col to int8
+        self.row = (
+            row_dict,
+            row_data[self.target_col].to_numpy(dtype=np.int8)[0],
+        )
 
     def _save_row(self) -> None:
         output_filename = f"{self.output_file_prefix}_{self.current_row_idx}.{self.file_extension}"  # noqa: E501
